+++ simple_history/tests/tests/test_models.py
@@ -10,9 +10,12 @@ from django.contrib.auth import get_user_model
 from django.core.files.base import ContentFile
 from django.db import models
 from django.db.models.fields.proxy import OrderWrt
-from django.test import TestCase
+from django.test import override_settings, TestCase
+from django.urls import reverse
 
 from simple_history.models import HistoricalRecords, ModelChange
+from simple_history.signals import pre_create_historical_record
+from simple_history.tests.tests.utils import middleware_override_settings
 from simple_history.utils import update_change_reason
 from ..external.models import ExternalModel2, ExternalModel4
 from ..models import (
@@ -31,6 +34,7 @@ from ..models import (
     Contact,
     ContactRegister,
     Country,
+    CustomNameModel,
     Document,
     Employee,
     ExternalModel1,
@@ -39,6 +43,7 @@ from ..models import (
     HistoricalChoice,
     HistoricalCustomFKError,
     HistoricalPoll,
+    HistoricalPollWithHistoricalIPAddress,
     HistoricalState,
     Library,
     MultiOneToOne,
@@ -48,6 +53,7 @@ from ..models import (
     PollInfo,
     PollWithExcludeFields,
     PollWithExcludedFKField,
+    PollWithHistoricalIPAddress,
     Province,
     Restaurant,
     SelfFK,
@@ -671,6 +677,19 @@ class CreateHistoryModelTests(unittest.TestCase):
                 "exception."
             )
 
+    def test_instantiate_history_model_with_custom_model_name(self):
+        try:
+            from ..models import MyHistoricalCustomNameModel
+        except ImportError:
+            self.fail("MyHistoricalCustomNameModel is in wrong module")
+        historical_model = MyHistoricalCustomNameModel()
+        self.assertEqual(
+            historical_model.__class__.__name__, "MyHistoricalCustomNameModel"
+        )
+        self.assertEqual(
+            historical_model._meta.db_table, "tests_myhistoricalcustomnamemodel"
+        )
+
 
 class AppLabelTest(TestCase):
     def get_table_name(self, manager):
@@ -1094,3 +1113,96 @@ class ExcludeForeignKeyTest(TestCase):
         historical = self.get_first_historical()
         instance = historical.instance
         self.assertEqual(instance.place, new_place)
+
+
+def add_static_history_ip_address(sender, **kwargs):
+    history_instance = kwargs["history_instance"]
+    history_instance.ip_address = "192.168.0.1"
+
+
+class ExtraFieldsStaticIPAddressTestCase(TestCase):
+    def setUp(self):
+        pre_create_historical_record.connect(
+            add_static_history_ip_address,
+            sender=HistoricalPollWithHistoricalIPAddress,
+            dispatch_uid="add_static_history_ip_address",
+        )
+
+    def tearDown(self):
+        pre_create_historical_record.disconnect(
+            add_static_history_ip_address,
+            sender=HistoricalPollWithHistoricalIPAddress,
+            dispatch_uid="add_static_history_ip_address",
+        )
+
+    def test_extra_ip_address_field_populated_on_save(self):
+        poll = PollWithHistoricalIPAddress.objects.create(
+            question="Will it blend?", pub_date=today
+        )
+
+        poll_history = poll.history.first()
+
+        self.assertEquals("192.168.0.1", poll_history.ip_address)
+
+    def test_extra_ip_address_field_not_present_on_poll(self):
+        poll = PollWithHistoricalIPAddress.objects.create(
+            question="Will it blend?", pub_date=today
+        )
+
+        with self.assertRaises(AttributeError):
+            poll.ip_address
+
+
+def add_dynamic_history_ip_address(sender, **kwargs):
+    history_instance = kwargs["history_instance"]
+    history_instance.ip_address = HistoricalRecords.thread.request.META["REMOTE_ADDR"]
+
+
+@override_settings(**middleware_override_settings)
+class ExtraFieldsDynamicIPAddressTestCase(TestCase):
+    def setUp(self):
+        pre_create_historical_record.connect(
+            add_dynamic_history_ip_address,
+            sender=HistoricalPollWithHistoricalIPAddress,
+            dispatch_uid="add_dynamic_history_ip_address",
+        )
+
+    def tearDown(self):
+        pre_create_historical_record.disconnect(
+            add_dynamic_history_ip_address,
+            sender=HistoricalPollWithHistoricalIPAddress,
+            dispatch_uid="add_dynamic_history_ip_address",
+        )
+
+    def test_signal_is_able_to_retrieve_request_from_thread(self):
+        data = {"question": "Will it blend?", "pub_date": "2018-10-30"}
+
+        self.client.post(reverse("pollip-add"), data=data)
+
+        polls = PollWithHistoricalIPAddress.objects.all()
+        self.assertEqual(1, polls.count())
+
+        poll_history = polls[0].history.first()
+        self.assertEqual("127.0.0.1", poll_history.ip_address)
+
+
+class WarningOnAbstractModelWithInheritFalseTest(TestCase):
+    def test_warning_on_abstract_model_with_inherit_false(self):
+
+        with warnings.catch_warnings(record=True) as w:
+
+            class AbstractModelWithInheritFalse(models.Model):
+                string = models.CharField()
+                history = HistoricalRecords()
+
+                class Meta:
+                    abstract = True
+
+            self.assertEqual(len(w), 1)
+            self.assertTrue(issubclass(w[0].category, UserWarning))
+            self.assertEqual(
+                str(w[0].message),
+                "HistoricalRecords added to abstract model "
+                "(AbstractModelWithInheritFalse) without "
+                "inherit=True",
+            )