+++ CHANGES.rst
@@ -2,9 +2,18 @@ Changes
 =======
 
 Unreleased
------------
+----------
+- Add `custom_model_name` parameter to the constructor of `HistoricalRecords` (gh-451)
+- Fix header on history pages when custom site_header is used (gh-448)
+- Modify `pre_create_historircal_record` to pass `history_instance` for ease of customization (gh-421)
+- Raise warning if HistoricalRecords(inherit=False) is in an abstract model (gh-341)
+
+2.5.1 (2018-10-19)
+------------------
 - Add `'+'` as the `history_type` for each instance in `bulk_history_create` (gh-449)
 - Add support for  `history_change_reason` for each instance in `bulk_history_create` (gh-449)
+- Add `history_change_reason` in the history list view under the  `Change reason` display name (gh-458)
+- Fix bug that caused failures when using a custom user model (gh-459)
 
 2.5.0 (2018-10-18)
 ------------------